requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
django==4.2.4
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
jsonschema==4.19.0; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
openapi-codec==1.3.2
packaging==23.1; python_version >= '3.7'
pytz==2023.3
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.30.2; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.9.2; python_version >= '3.8'
simplejson==3.19.1; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.7.1; python_version < '3.11'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.4; python_version >= '3.7'
